Respond to first password prompt only.
SSSD supports multiple authentication types.  When OTP is enabled, pam_sss.so will return two prompts.  Users can either enter password on the first prompt and OTP on the second prompt, or password+OTP on the first prompt and ignore the second prompt.  This commit is to support the latter method.